requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# LangChain + LangGraph
langchain==0.3.12
langchain-groq==0.2.4
langchain-core==0.3.25
langgraph==0.2.56

# Groq SDK
groq==0.14.0

# Supabase
supabase==2.10.0

# PDF/Document Processing
PyMuPDF==1.24.0
python-docx==1.1.0

# Data Validation
pydantic==2.10.0
pydantic-settings==2.7.0

# Text Processing
tiktoken==0.8.0

# Utilities
python-dotenv==1.0.0
httpx==0.27.0
tenacity==9.0.0
structlog==24.1.0

# Development
pytest==8.0.0
pytest-asyncio==0.24.0
